services/image-generator/requirements.txt
sqlalchemy==2.0.23
pymysql==1.1.1
pydantic-settings==2.5.0

